Use vocdoni react providers
