+++ package.json
@@ -21,7 +21,7 @@
   },
   "dependencies": {
     "@apollo/client": "^3.5.8",
-    "@aragon/ods": "^0.2.10",
+    "@aragon/ods": "^0.2.12",
     "@aragon/sdk-client": "^1.12.0-rc1",
     "@elastic/apm-rum-react": "^1.4.4",
     "@radix-ui/react-accordion": "^1.1.2",
@@ -31,7 +31,8 @@
     "@tiptap/pm": "^2.0.3",
     "@tiptap/react": "^2.0.3",
     "@tiptap/starter-kit": "^2.0.3",
-    "@vocdoni/sdk": "^0.0.18",
+    "@vocdoni/react-providers": "^0.1.7",
+    "@vocdoni/sdk": "^0.1.1",
     "@walletconnect/core": "^2.8.3",
     "@walletconnect/utils": "^2.8.3",
     "@walletconnect/web3wallet": "^1.8.2",
